requirements.txt
asgiref==3.4.1
autopep8==1.6.0
black==21.8b0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.45.0
django-ckeditor==6.3.2
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==2.0.0
env==0.1.0
filelock==3.6.0
gunicorn==20.1.0
idna==3.2
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
virtualenv==20.13.3
whitenoise==5.3.0
